client/requirements.txt
pygame~=2.5.2
pyscroll~=2.31
PyTMX~=3.32
pyinstaller~=6.3.0
Pillow~=10.1.0